Change Int to Long

+++ 2024/Common.kt
@@ -1,5 +1,5 @@
 import java.io.File
 
-fun readInput(filename: String): List<List<Int>> = File(filename).readLines()
+fun readInput(filename: String): List<List<Long>> = File(filename).readLines()
     .map { Regex("(\\d+)").findAll(it).toList() }
-    .map { row -> row.map { it.value.toInt() } }
+    .map { row -> row.map { it.value.toLong() } }
